core/langgraph_design/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dev-cycle-design-graph"
version = "0.1.0"
description = "LangGraph orchestration for dev-cycle workflows (/design, /init-dev-cycle): guardrails + human checkpoints."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "langgraph>=1.0,<2",
  "GitPython>=3.1,<4",
  "PyYAML>=6,<7",
]

[project.scripts]
dev-cycle-design-graph = "design_graph.cli:main"
dev-cycle-init-graph = "init_graph.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["common", "design_graph", "init_graph", "queue_graph"]

[tool.hatch.build.targets.wheel.force-include]
"init_graph/prompt.md" = "init_graph/prompt.md"

[tool.hatch.build.targets.sdist]
include = ["/common", "/design_graph", "/init_graph", "/queue_graph"]
